add conmmand
