+++ test/outputs/config/get-current-user-regex.csv
@@ -1,0 +1,2 @@
+email,firstName,lastName,userId
+".+",".+",".+",".+"
