+++ test/outputs/config/get-current-user-regex.json
@@ -1,0 +1,6 @@
+\{
+  "userId": ".+",
+  "firstName": ".+",
+  "lastName": ".+",
+  "email": ".+"
+\}
