+++ test/outputs/config/get-current-user-regex.yaml
@@ -1,0 +1,4 @@
+userId: .+
+firstName: .+
+lastName: .+
+email: .+
