+++ test/test.bash
@@ -1,0 +1,5 @@
+if [[ "userId: 123" =~ "userId: .+" ]]; then
+  echo "Match found"
+else
+  echo "No match"
+fi
